+++ src/components/account/DeleteUserComponent.jsx
@@ -1,0 +1,10 @@
+import React, { Component } from "react";
+
+class DeleteUser extends Component {
+  state = {};
+  render() {
+    return <div></div>;
+  }
+}
+
+export default DeleteUser;
